+++ .gitignore
@@ -2,5 +2,4 @@
 .DS_Store
 node_modules/
 npm-debug.log
-.nyc_output
 coverage
