+++ .travis.yml
@@ -9,5 +9,4 @@ script:
   - npm run test:coverage -s
   - npm run test:codestyle -s
 after_success:
-  - npm run report:lcov
   - bash <(curl -s https://codecov.io/bash)
