+++ jestCoverage.json
@@ -1,0 +1,13 @@
+{
+  "automock": false,
+  "coverageDirectory": "coverage",
+  "collectCoverage": true,
+  "coverageThreshold": {
+    "global": {
+      "branches": 90,
+      "functions": 90,
+      "lines": 90,
+      "statements": 90
+    }
+  }
+}
